Ajustes varios y reorganizacion de la planificacion y pagina de inicio

+++ Xiloa/src/service/IService.java
@@ -347,7 +347,7 @@ public interface IService {
 	 * 
 	 */
 	
-	public List<Requisito> getRequisitos(int certificacionId);
+	public List<Requisito> getRequisitos(Long certificacionId);
 	
 	/*
 	 * @param el requisito a guardar
